OneDrive/Desktop/githublocal/requirements.txt
discord.py
fastapi
uvicorn
asyncpg
python-dotenv
requests
pyngrok
openai
nest-asyncio
